refactor: split fast_io lib.rs into focused submodules
Decompose the 1555-line crates/fast_io/src/lib.rs facade into four
purpose-built submodules so the crate root reads as a navigable index
of re-exports rather than mixing crate docs with inline impls and
tests:

- policy.rs: IoUringPolicy, CowPolicy, IocpPolicy, ZeroCopyPolicy
- io_uring_depth.rs: IO_URING_DEPTH_{MIN,MAX}, IoUringDepthError,
  validate_io_uring_depth
- io_uring_ops.rs: hard_link, try_hard_link_via_io_uring,
  try_rename_via_io_uring, try_statx_batch_via_io_uring
- status.rs: iocp_status_detail, io_uring_status_detail,
  io_uring_availability_reason, io_uring_kernel_info,
  platform_io_capabilities

Each submodule carries its own #[cfg(test)] block. Public paths are
preserved verbatim: every fast_io::Item that compiled before still
resolves to the same item. lib.rs shrinks from 1555 to 219 lines.
